+++ controllers/productsController.js
@@ -4,14 +4,17 @@ import Review from '../model/reviewModel.js';
 
 //add a product
 export const addNew = async(req,res) => {
-
     const asin = req.body.asin.toUpperCase();
     const { category, product, type, price, reviews, title, thumbnail, description, images, dimensions, weight, manufacturer, model_number, sold_by, brand } = req.body;
     const newProduct = new Products({ asin, category, product, type, price, reviews, title, thumbnail });
-    const features = req.body.features.split(',');
+    const features = req.body.features?req.body.features.split(','):[""];
 
     const date = new Date();
     const month = date.toLocaleString('default', { month: 'long' });
+
+    if(!req.session.user && req.session.user.role !=='Admin') {	
+        return res.status(400).send('No Session Found! Please Login Again')	
+    }
 
     let data = {
         asin: asin,
@@ -20,12 +23,13 @@ export const addNew = async(req,res) => {
         feature_bullets: features,
         reviews: reviews,
         price: price,
+        main_image: thumbnail,
         total_images: images.length,
         images: images,
         product_information: {
             dimensions: dimensions,
             weight: weight,
-            available_from: `${date.getDay()} ${month} ${date.getFullYear()}`,
+            available_from: `${date.getDate()} ${month} ${date.getFullYear()}`,
             available_from_utc: date.toISOString(),
             manufacturer: manufacturer,
             model_number: model_number,
@@ -63,6 +67,11 @@ export const addNew = async(req,res) => {
 
 //add multiple products
 export const addMultiple = async(req,res) => {
+
+    if(!req.session.user && req.session.user.role !=='Admin') {	
+        return res.status(400).send('No Session Found! Please Login Again')	
+    }
+
     const data = req.body.products
     
     try {       
@@ -79,6 +88,11 @@ export const addMultiple = async(req,res) => {
 //get all products
 export const getProducts = async(req,res) => {
     try{
+
+        if(!req.session.user) {	
+            return res.status(400).send('No Session Found! Please Login Again')	
+        }	
+
         const result = await Products.find({})
         if(result.length <1) return res.status(404).send({"err":"No Data Found"});
         res.status(200).send(result)
@@ -91,6 +105,11 @@ export const getProducts = async(req,res) => {
 //get products by category or (product and type) or asin
 export const getProdsByCategory = async(req,res) => {
     try{
+
+        // if(!req.session.user) {	
+        //     return res.status(400).send('No Session Found! Please Login Again')	
+        // }
+
         const category = req.query.category;
         const product = req.query.product;
         const type = req.query.type;
@@ -123,6 +142,11 @@ export const getProdsByCategory = async(req,res) => {
 //delete product 
 export const deleteProduct = async (req, res) => {
     try {
+
+        if(!req.session.user && req.session.user.role !=='Admin') {	
+            return res.status(400).send('No Session Found! Please Login Again')	
+        }
+
         await Products.findByIdAndDelete(req.params.id);
 
         return res.status.send({"success": "Product deleted successfully"})
